+++ internal/infrastructure/persistence/project/fetch/localStorage.go
@@ -3,9 +3,6 @@ package fetch
 import (
 	"errors"
 	"fmt"
-	"io"
-	"os"
-	"path/filepath"
 
 	"github.com/apenella/ransidble/internal/domain/core/entity"
 	"github.com/apenella/ransidble/internal/domain/ports/repository"
@@ -20,17 +17,12 @@ const (
 var (
 	// ErrProjectNotProvided represents an error when the project is not provided
 	ErrProjectNotProvided = errors.New("project not provided")
-	// ErrCopyingFilesInWorkingDir represents an error copying files in working directory
-	ErrCopyingFilesInWorkingDir = errors.New("error copying files in working directory")
-	// ErrWorkingDirNotExists represents an error when the destination to fetch does not exists
-	ErrWorkingDirNotExists = errors.New("working directory does not exists")
 )
 
 // LocalStorage represents a repository on local storage
 type LocalStorage struct {
 	// Filesystem path where projects are stored
-	Fs afero.Fs
-
+	fs afero.Fs
 	// logger is the logger
 	logger repository.Logger
 }
@@ -38,15 +30,15 @@ type LocalStorage struct {
 // NewLocalStorage creates a new local project repository
 func NewLocalStorage(fs afero.Fs, logger repository.Logger) *LocalStorage {
 	return &LocalStorage{
-		Fs:     fs,
+		fs:     fs,
 		logger: logger,
 	}
 }
 
 // Fetch method copies the project from local storage to working directory
-func (s *LocalStorage) Fetch(project *entity.Project, workingDir string) error {
+func (s *LocalStorage) Fetch(project *entity.Project, workingDir string) (err error) {
 
-	var err error
+	var sourceCodeFetcher SourceCodeFetcher
 	var workingDirExist bool
 
 	if project == nil {
@@ -59,9 +51,9 @@ func (s *LocalStorage) Fetch(project *entity.Project, workingDir string) error {
 		return ErrProjectNotProvided
 	}
 
-	_, err = s.Fs.Stat(workingDir)
+	_, err = s.fs.Stat(workingDir)
 	if err != nil {
-		workingDirExist, err = afero.DirExists(s.Fs, workingDir)
+		workingDirExist, err = afero.DirExists(s.fs, workingDir)
 		if workingDirExist == false || err != nil {
 			s.logger.Error(
 				ErrWorkingDirNotExists.Error(),
@@ -81,107 +73,33 @@ func (s *LocalStorage) Fetch(project *entity.Project, workingDir string) error {
 		"working_dir": workingDir,
 	})
 
-	err = afero.Walk(s.Fs, project.Reference, func(path string, info os.FileInfo, err error) error {
-
-		if err != nil {
-			errorMsg := fmt.Sprintf("error walking through %s: %s", path, err)
-			s.logger.Error(
-				errorMsg,
-				map[string]interface{}{
-					"component": "LocalStorage.Fetch",
-					"package":   "github.com/apenella/ransidble/internal/infrastructure/persistence/project/fetch",
-				})
-			return fmt.Errorf("%s", errorMsg)
-		}
-
-		relPath, err := filepath.Rel(project.Reference, path)
-		if err != nil {
-			errorMsg := fmt.Sprintf("error getting relative path for %s: %s", path, err)
-			s.logger.Error(
-				errorMsg,
-				map[string]interface{}{
-					"component": "LocalStorage.Fetch",
-					"package":   "github.com/apenella/ransidble/internal/infrastructure/persistence/project/fetch",
-				})
-			return fmt.Errorf("%s", errorMsg)
-		}
-
-		if project.Reference == path {
-			return nil
-		}
-
-		if info.IsDir() {
-			err = s.Fs.MkdirAll(filepath.Join(workingDir, relPath), 0755)
-			if err != nil {
-				errorMsg := fmt.Sprintf("error creating directory %s: %s", filepath.Join(workingDir, relPath), err)
-				s.logger.Error(
-					errorMsg,
-					map[string]interface{}{
-						"component": "LocalStorage.Fetch",
-						"package":   "github.com/apenella/ransidble/internal/infrastructure/persistence/project/fetch",
-					})
-
-				return fmt.Errorf("%s", errorMsg)
-
-			}
-		}
-
-		srcFile, err := s.Fs.Open(path)
-		if err != nil {
-			errorMsg := fmt.Sprintf("error opening file %s: %s", path, err)
-			s.logger.Error(
-				errorMsg,
-				map[string]interface{}{
-					"component": "LocalStorage.Fetch",
-					"package":   "github.com/apenella/ransidble/internal/infrastructure/persistence/project/fetch",
-				})
-
-			return fmt.Errorf("%s", errorMsg)
-		}
-		defer srcFile.Close()
-
-		destPath := filepath.Join(workingDir, relPath)
-		destFile, err := s.Fs.Create(destPath)
-		if err != nil {
-			errorMsg := fmt.Sprintf("error creating file %s: %s", destPath, err)
-			s.logger.Error(
-				errorMsg,
-				map[string]interface{}{
-					"component": "LocalStorage.Fetch",
-					"package":   "github.com/apenella/ransidble/internal/infrastructure/persistence/project/fetch",
-				})
-
-			return fmt.Errorf("%s", errorMsg)
-		}
-		defer destFile.Close()
-
-		_, err = io.Copy(destFile, srcFile)
-		if err != nil {
-			errorMsg := fmt.Sprintf("%s: From %s to %s: %s", ErrCopyingFilesInWorkingDir, path, destPath, err)
-			s.logger.Error(
-				errorMsg,
-				map[string]interface{}{
-					"component": "LocalStorage.Fetch",
-					"package":   "github.com/apenella/ransidble/internal/infrastructure/persistence/project/fetch",
-				})
-
-			return fmt.Errorf("%s", errorMsg)
-		}
-
-		return nil
-
-	})
-
+	infoProjectReference, err := s.fs.Stat(project.Reference)
 	if err != nil {
-		errorMsg := fmt.Sprintf("%s: %s", ErrCopyingFilesInWorkingDir, err)
 		s.logger.Error(
-			errorMsg,
+			err.Error(),
+			map[string]interface{}{
+				"component": "LocalStorage.Fetch",
+				"package":   "github.com/apenella/ransidble/internal/infrastructure/persistence/project/fetch",
+			})
+		return fmt.Errorf("%w", err)
+	}
+
+	if infoProjectReference.IsDir() {
+		sourceCodeFetcher = NewLocalFetchDir(s.fs, s.logger)
+	} else {
+		sourceCodeFetcher = NewLocalFetchFile(s.fs, s.logger)
+	}
+
+	err = sourceCodeFetcher.Fetch(project.Reference, workingDir)
+	if err != nil {
+		s.logger.Error(
+			err.Error(),
 			map[string]interface{}{
 				"component": "LocalStorage.Fetch",
 				"package":   "github.com/apenella/ransidble/internal/infrastructure/persistence/project/fetch",
 			})
 
-		return fmt.Errorf("%s", errorMsg)
+		return fmt.Errorf("%w", err)
 	}
 
 	return nil