+++ RELEASE_NOTES.md
@@ -4,7 +4,10 @@
 
 ### Added
 
+- Use the local filesystem to store project files
+- Define a `plain` project format, when the project is stored in the local filesystem
+- Define a `tar.gz` project format, when the project is stored in the local filesystem
 - Rest API endpoint to create a task to execute an Ansible playbook command 
-- Rest API endpoint to get the status of a task
 - Rest API endpoint to get a list of all projects
 - Rest API endpoint to get project details
+- Rest API endpoint to get the status of a task